update: componentize
