+++ packages/dashboard/components/consent/utils/consentArrayUtils.ts
@@ -1,0 +1,348 @@
+import { VersionParameter } from '../types';
+import { isEmptyParameterValue, decodeParameterValue } from './parameterDecoding';
+import { encodeParameterValue } from '../../../utils/parameterEncoding';
+
+/**
+ * Prepares parameter removal data for contract calls
+ * Creates the required array structure for removeToolPolicyParameters
+ */
+export const prepareParameterRemovalData = (
+  parametersToRemove: VersionParameter[],
+  versionInfo: any
+) => {
+  const removalToolIpfsCids: string[] = [];
+  const removalPolicyIpfsCids: string[][] = [];
+  const removalParameterNames: string[][][] = [];
+
+  parametersToRemove.forEach((param) => {
+    const { toolIndex, policyIndex, name: paramName } = param;
+
+    // Ensure arrays exist at this level
+    while (removalToolIpfsCids.length <= toolIndex) {
+      removalToolIpfsCids.push('');
+      removalPolicyIpfsCids.push([]);
+      removalParameterNames.push([]);
+    }
+
+    // Set the tool IPFS CID if not already set
+    if (!removalToolIpfsCids[toolIndex] && versionInfo) {
+      const toolsData = versionInfo.appVersion?.tools || versionInfo[1]?.[3];
+      if (toolsData && toolsData[toolIndex] && toolsData[toolIndex][0]) {
+        removalToolIpfsCids[toolIndex] = toolsData[toolIndex][0];
+      }
+    }
+
+    // Ensure policy arrays exist
+    while (removalPolicyIpfsCids[toolIndex].length <= policyIndex) {
+      removalPolicyIpfsCids[toolIndex].push('');
+      removalParameterNames[toolIndex].push([]);
+    }
+
+    // Set the policy IPFS CID
+    if (!removalPolicyIpfsCids[toolIndex][policyIndex] && versionInfo) {
+      const toolsData = versionInfo.appVersion?.tools || versionInfo[1]?.[3];
+      if (
+        toolsData &&
+        toolsData[toolIndex] &&
+        toolsData[toolIndex][1] &&
+        toolsData[toolIndex][1][policyIndex] &&
+        toolsData[toolIndex][1][policyIndex][0]
+      ) {
+        removalPolicyIpfsCids[toolIndex][policyIndex] =
+          toolsData[toolIndex][1][policyIndex][0];
+      }
+    }
+
+    // Add the parameter name
+    if (removalPolicyIpfsCids[toolIndex][policyIndex]) {
+      removalParameterNames[toolIndex][policyIndex].push(paramName);
+    }
+  });
+
+  // Filter out empty tool entries
+  const filteredToolIndices = removalToolIpfsCids
+    .map((_, i) => i)
+    .filter((i) => removalToolIpfsCids[i] !== '');
+
+  // Create final arrays for the contract call
+  const filteredTools: string[] = [];
+  const filteredPolicies: string[][] = [];
+  const filteredParams: string[][][] = [];
+
+  // Process each valid tool
+  filteredToolIndices.forEach((toolIndex) => {
+    // Find valid policies for this tool
+    const validPolicyIndices = removalPolicyIpfsCids[toolIndex]
+      .map((_, i) => i)
+      .filter(
+        (i) =>
+          removalPolicyIpfsCids[toolIndex][i] !== '' &&
+          removalParameterNames[toolIndex][i].length > 0,
+      );
+
+    if (validPolicyIndices.length > 0) {
+      filteredTools.push(removalToolIpfsCids[toolIndex]);
+
+      // Gather policies and params
+      const toolPolicies: string[] = [];
+      const toolParams: string[][] = [];
+
+      validPolicyIndices.forEach((policyIndex) => {
+        toolPolicies.push(removalPolicyIpfsCids[toolIndex][policyIndex]);
+        toolParams.push(removalParameterNames[toolIndex][policyIndex]);
+      });
+
+      filteredPolicies.push(toolPolicies);
+      filteredParams.push(toolParams);
+    }
+  });
+
+  return {
+    filteredTools,
+    filteredPolicies,
+    filteredParams,
+  };
+};
+
+/**
+ * Prepares parameter update data for contract calls
+ * Creates the required array structure for setToolPolicyParameters
+ */
+export const prepareParameterUpdateData = (
+  parameters: VersionParameter[],
+  versionInfo: any
+) => {
+  const toolIpfsCids: string[] = [];
+  const policyIpfsCids: string[][] = [];
+  const policyParameterNames: string[][][] = [];
+  const policyParameterValues: Uint8Array[][][] = [];
+
+  if (!versionInfo) {
+    return { 
+      toolIpfsCids, 
+      policyIpfsCids, 
+      policyParameterNames, 
+      policyParameterValues,
+      hasParametersToSet: false 
+    };
+  }
+
+  const toolsData = versionInfo.appVersion?.tools || versionInfo[1]?.[3];
+
+  if (toolsData && Array.isArray(toolsData)) {
+    toolsData.forEach((tool: any, toolIndex: number) => {
+      if (!tool || !Array.isArray(tool)) return;
+
+      const toolIpfsCid = tool[0];
+      if (toolIpfsCid) {
+        toolIpfsCids[toolIndex] = toolIpfsCid;
+        policyIpfsCids[toolIndex] = [];
+        policyParameterNames[toolIndex] = [];
+        policyParameterValues[toolIndex] = [];
+
+        const policies = tool[1];
+        if (Array.isArray(policies)) {
+          policies.forEach((policy: any, policyIndex: number) => {
+            if (!policy || !Array.isArray(policy)) return;
+
+            const policyIpfsCid = policy[0];
+            policyIpfsCids[toolIndex][policyIndex] = policyIpfsCid;
+            policyParameterNames[toolIndex][policyIndex] = [];
+            policyParameterValues[toolIndex][policyIndex] = [];
+
+            const paramNames = policy[1];
+
+            if (Array.isArray(paramNames)) {
+              // Filter parameters that have user-provided values
+              paramNames.forEach((name: any, paramIndex: number) => {
+                // Find matching parameter value from user input
+                const param = parameters.find(
+                  (p) =>
+                    p.toolIndex === toolIndex &&
+                    p.policyIndex === policyIndex &&
+                    p.paramIndex === paramIndex,
+                );
+
+                // Only add parameters that have user-provided values and aren't empty
+                if (param && param.value !== undefined) {
+                  // Check if parameter is empty using the shared utility
+                  const isEmpty = isEmptyParameterValue(param.value, param.type);
+
+                  // Skip if parameter is empty
+                  if (isEmpty) return;
+
+                  const paramName =
+                    typeof name === 'string' && name.trim() !== ''
+                      ? name.trim()
+                      : `param_${paramIndex}`;
+
+                  policyParameterNames[toolIndex][policyIndex].push(paramName);
+                  policyParameterValues[toolIndex][policyIndex].push(
+                    encodeParameterValue(param.type, param.value, paramName),
+                  );
+                }
+              });
+            }
+          });
+        }
+      }
+    });
+  }
+
+  // Check if there are any parameters to set
+  const hasParametersToSet = toolIpfsCids.some((toolCid, toolIndex) => {
+    if (policyIpfsCids[toolIndex]) {
+      return policyIpfsCids[toolIndex].some((policyCid, policyIndex) => {
+        if (
+          policyParameterNames[toolIndex] &&
+          policyParameterNames[toolIndex][policyIndex]
+        ) {
+          return policyParameterNames[toolIndex][policyIndex].length > 0;
+        }
+        return false;
+      });
+    }
+    return false;
+  });
+
+  return {
+    toolIpfsCids,
+    policyIpfsCids,
+    policyParameterNames,
+    policyParameterValues,
+    hasParametersToSet
+  };
+};
+
+/**
+ * Identify parameters that need to be removed based on empty values
+ */
+export const identifyParametersToRemove = (
+  existingParameters: VersionParameter[],
+  parameters: VersionParameter[]
+) => {
+  const parametersToRemove: VersionParameter[] = [];
+  
+  existingParameters.forEach((existingParam) => {
+    try {
+      // Make sure we have a decoded value for comparison
+      if (typeof existingParam.value === 'string' && existingParam.type) {
+        existingParam.value = decodeParameterValue(existingParam.value, existingParam.type);
+      }
+      
+      // Match by name - the most direct approach
+      const formParam = parameters.find((p) => p.name === existingParam.name);
+
+      // If no matching parameter in form, remove it
+      if (!formParam) {
+        console.log(`Parameter ${existingParam.name} not found in form, will be removed`);
+        parametersToRemove.push(existingParam);
+      }
+      // If parameter exists but is empty/default, remove if existing value is not empty
+      else if (isEmptyParameterValue(formParam.value, formParam.type)) {
+        // Only remove if the existing value wasn't also empty/zero
+        if (!isEmptyParameterValue(existingParam.value, existingParam.type)) {
+          console.log(`Parameter ${existingParam.name} will be removed: value changed to empty`);
+          parametersToRemove.push(existingParam);
+        }
+      }
+    } catch (error) {
+      console.error(`Error checking parameter ${existingParam.name} for removal:`, error);
+    }
+  });
+
+  return parametersToRemove;
+};
+
+/**
+ * Prepares version permit data from versionInfo for consent approval
+ * 
+ * @param versionInfo - The version information containing tools, policies, and parameters
+ * @param parameters - User-provided parameter values
+ * @returns Formatted arrays for contract call (toolIpfsCids, toolPolicies, toolPolicyParameterNames, toolPolicyParameterTypes)
+ */
+export const prepareVersionPermitData = (
+  versionInfo: any,
+  parameters: any[]
+) => {
+  const toolIpfsCids: string[] = [];
+  const toolPolicies: string[][] = [];
+  const toolPolicyParameterNames: string[][][] = [];
+  const toolPolicyParameterTypes: number[][][] = [];
+
+  if (versionInfo) {
+    const toolsData = versionInfo.appVersion?.tools || versionInfo[1]?.[3];
+
+    if (toolsData && Array.isArray(toolsData)) {
+      toolsData.forEach((tool: any, toolIndex: number) => {
+        if (!tool || !Array.isArray(tool)) return;
+
+        const toolIpfsCid = tool[0];
+        if (toolIpfsCid) {
+          toolIpfsCids[toolIndex] = toolIpfsCid;
+        }
+
+        toolPolicies[toolIndex] = [];
+        toolPolicyParameterNames[toolIndex] = [];
+        toolPolicyParameterTypes[toolIndex] = [];
+
+        const policies = tool[1];
+        if (Array.isArray(policies)) {
+          policies.forEach((policy: any, policyIndex: number) => {
+            if (!policy || !Array.isArray(policy)) return;
+
+            toolPolicies[toolIndex][policyIndex] = policy[0];
+            toolPolicyParameterNames[toolIndex][policyIndex] = [];
+            toolPolicyParameterTypes[toolIndex][policyIndex] = [];
+
+            // Extract the actual parameter names and types from the policy
+            const paramNames = policy[1];
+            const paramTypes = policy[2];
+
+            if (Array.isArray(paramNames) && Array.isArray(paramTypes)) {
+              // Use the actual parameter names from the version info
+              paramNames.forEach((name: any, paramIndex: number) => {
+                // Ensure parameter name is never empty by using a default if it's empty
+                const paramName =
+                  typeof name === 'string' && name.trim() !== ''
+                    ? name.trim()
+                    : `param_${paramIndex}`;
+
+                toolPolicyParameterNames[toolIndex][policyIndex][paramIndex] = paramName;
+
+                // Set the parameter type if available
+                if (paramTypes[paramIndex] !== undefined) {
+                  toolPolicyParameterTypes[toolIndex][policyIndex][paramIndex] =
+                    typeof paramTypes[paramIndex] === 'number'
+                      ? paramTypes[paramIndex]
+                      : 0;
+                } else {
+                  toolPolicyParameterTypes[toolIndex][policyIndex][paramIndex] = 0;
+                }
+              });
+            }
+          });
+        }
+      });
+    }
+  }
+
+  // Apply user-provided parameter types
+  if (parameters.length > 0) {
+    parameters.forEach((param) => {
+      if (
+        toolPolicyParameterTypes[param.toolIndex] &&
+        toolPolicyParameterTypes[param.toolIndex][param.policyIndex]
+      ) {
+        toolPolicyParameterTypes[param.toolIndex][param.policyIndex][param.paramIndex] = param.type;
+      }
+    });
+  }
+
+  return {
+    toolIpfsCids,
+    toolPolicies,
+    toolPolicyParameterNames,
+    toolPolicyParameterTypes
+  };
+};
